# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMFACA Crédito API",
        "description": "API para la gestión de solicitudes de crédito de COMFACA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/lineas_credito/parametros": {
            "get": {
                "tags": [
                    "LineasCredito"
                ],
                "summary": "Obtener parámetros generales",
                "description": "Obtener parámetros generales de líneas de crédito",
                "operationId": "e1c54dd0db67693bd7f69762cd9bdb1d",
                "responses": {
                    "200": {
                        "description": "Parámetros obtenidos"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas_credito/tipo_creditos": {
            "get": {
                "tags": [
                    "LineasCredito"
                ],
                "summary": "Obtener tipos de crédito",
                "description": "Obtener tipos de crédito disponibles",
                "operationId": "d26db0d38e035ad8636cfb53cc71d2ac",
                "responses": {
                    "200": {
                        "description": "Tipos de crédito obtenidos"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas_credito/completo": {
            "get": {
                "tags": [
                    "LineasCredito"
                ],
                "summary": "Obtener líneas de crédito completas",
                "description": "Obtener información completa de líneas de crédito (parámetros + tipos)",
                "operationId": "1fe68864331882e27d5e3448c61768c1",
                "responses": {
                    "200": {
                        "description": "Líneas completas obtenidas"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-credito/disponibilidad": {
            "get": {
                "tags": [
                    "LineasCredito"
                ],
                "summary": "Verificar disponibilidad del servicio",
                "description": "Verificar disponibilidad del servicio externo",
                "operationId": "9f92a5e5eb7d6146c94b1c921b8d4e10",
                "responses": {
                    "200": {
                        "description": "Servicio disponible"
                    },
                    "503": {
                        "description": "Servicio no disponible"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-credito/estadisticas": {
            "get": {
                "tags": [
                    "LineasCredito"
                ],
                "summary": "Obtener estadísticas de líneas de crédito",
                "description": "Obtener estadísticas de líneas de crédito",
                "operationId": "0bc2efe2d12a74be886a00084b7a543e",
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}